% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficarPunto.r
\name{graficarPunto}
\alias{graficarPunto}
\title{graficarPunto
Grafica un punto dado}
\usage{
graficarPunto(a)
}
\arguments{
\item{a}{vector}
}
\value{
grafico
}
\description{
graficarPunto
Grafica un punto dado
}
\examples{
graficarPunto(c(1,1))

}
