% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficarSegemento.r
\name{graficarSegmento}
\alias{graficarSegmento}
\title{graficarSegmento()
Grafica un segmento a partir de 2 puntos dados}
\usage{
graficarSegmento(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
grafica
}
\description{
graficarSegmento()
Grafica un segmento a partir de 2 puntos dados
}
\examples{
graficarSegmento(c(1,1),c(2,2))


}
