% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntomedio.r
\name{puntomedio}
\alias{puntomedio}
\title{puntomedio() calcula el punto medio entre 2 puntos}
\usage{
puntomedio(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
vector
}
\description{
puntomedio() calcula el punto medio entre 2 puntos
}
\examples{
puntomedio(c(1,1),c(2,2))
# [1] 1,5 1,5

}
